% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhdynamics-package.R
\docType{package}
\name{hhdynamics-package}
\alias{hhdynamics}
\alias{hhdynamics-package}
\title{hhdynamics: Fitting Household Transmission Model to Estimate Household Transmission Dynamics of Influenza}
\description{
A Bayesian household transmission model to estimate household transmission dynamics, with accounting for infection from community and tertiary cases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timktsang/hhdynamics}
  \item Report bugs at \url{https://github.com/timktsang/hhdynamics/issues}
}

}
\author{
\strong{Maintainer}: Tim Tsang \email{timkltsang@gmail.com} (\href{https://orcid.org/0000-0001-5037-6776}{ORCID})

}
\examples{
\donttest{
# Fit a household transmission model
data(inputdata)
fit <- household_dynamics(inputdata, inf_factor = ~sex, sus_factor = ~age)
summary(fit)
}
}
\keyword{internal}
