% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{inputdata}
\alias{inputdata}
\title{Example of input data}
\format{
A example data with 8 variables, where each row represents an individual. For user's data, more predictors could be added by adding columns. For categorical variables, it should be specificed as factor, by using as.factor() function. The date is the dataset is an integer, by selecting a reference date as day 1. In the example dataset, 2008-01-01 is day 1.
\describe{
  \item{hhID}{The household id}	
  \item{member}{The member id in a households, use 0 to index, 1 and so on as household contact}	
  \item{size}{The number of individuals in the household}
  \item{end}{The end date of follow-up of that individual}
  \item{inf}{The infection status of the member. By defintion, index must be infected}	
  \item{onset}{The onset time of infected individual}	
  \item{age_group}{The age group of individual. 0: 0-19, 1: 20-64, 2: 65 or above}
  \item{sex}{The sex of the individual. 0: Female, 1: Male}
}
}
\usage{
data(inputdata)
}
\description{
This is an example of the input data used in the \code{hhdynamics} function. This data frame illustrates the format of the input data. This is a simulated data.
}
\concept{inputdata}
\keyword{datasets}
