% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{para}
\alias{para}
\title{Example of parameter vector for the main model}
\format{
A vector with 8 elements, where each of them is a model parameter:
\describe{
  \item{element 1}{the random effect of individual infectivity (not available in this version. Models assumed no indiviudal heterogeneity after accounting for factors affecting infectivity)}
  \item{element 2}{the probability of infection from the community}
  \item{element 3}{the probability of person-to-person transmission in households}
  \item{element 4}{the parameter of the relationship between the number of household contacts and transmission (default 0 in this version)}
  \item{element 5}{the parameter for infectivity of male (Reference group: male)}
  \item{element 6}{the relative susceptibility of age group 1 (Reference group: age group 0)}
  \item{element 7}{the relative susceptibility of age group 2 (Reference group: age group 0)}
}
}
\usage{
data(para)
}
\description{
This is an example of the parameter vector for the main model used in the \code{hhdynamics} function. This vector specifies the format of the parameter vector for the main model.
}
\seealso{
Other example_data: 
\code{\link{SI}}
}
\concept{example_data}
\keyword{datasets}
