% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhdynamics_fit.R
\name{print.summary.hhdynamics_fit}
\alias{print.summary.hhdynamics_fit}
\title{Print method for summary.hhdynamics_fit}
\usage{
\method{print}{summary.hhdynamics_fit}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{summary.hhdynamics_fit}.}

\item{digits}{Number of significant digits for printing. Default is 3.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for summary.hhdynamics_fit
}
