% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{SI}
\alias{SI}
\title{Example of serial interval distribution of influenza}
\format{
A vector of length 14, where element X of the vector represents a probability that the onset day is the X days after the onset day of the infector.
\describe{
  \item{}{This is the serial interval distribution. The sum of the vector elements should be 1.}
}
}
\usage{
data(SI)
}
\description{
This is an example of the serial interval distribution used in the \code{hhdynamics} function. This vector specifies the format of the serial interval distribution. This is estimated from Tsang et al. Association between antibody titers and protection against influenza virus infection within households. J Infect Dis. 2014 Sep 1;210(5):684-92.
}
\seealso{
Other example_data: 
\code{\link{para}}
}
\concept{example_data}
\keyword{datasets}
