% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_function.R
\name{create_wide_data}
\alias{create_wide_data}
\title{Create a wide format of household data}
\usage{
create_wide_data(input, inf_factor, sus_factor)
}
\arguments{
\item{input}{The input data, in long format.}

\item{inf_factor}{Factors affecting infectivity. Use the format of '~ factor1 + factor2'.}

\item{sus_factor}{Factors affecting susceptibility. Use the format of '~ factor1 + factor2'.}
}
\value{
A data frame in the wide format.
}
\description{
This function creates a wide format of the household data (each row is a household), based on the long format of data (each row is an individual), to fit the household transmission model.
}
\examples{
wide_data <- create_wide_data(inputdata,'~sex','~age')
}
