% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_function.R
\name{household_dynamics}
\alias{household_dynamics}
\title{Fitting household transmission model to the data, based on Markov chain Monte Carlo (MCMC).}
\usage{
household_dynamics(
  input,
  inf_factor,
  sus_factor,
  SI,
  n_iteration = 15000,
  burnin = 5000,
  thinning = 1
)
}
\arguments{
\item{input}{The input data, in long format (each row is an individual). The data must include: 1) hhID: the household id, 2) member: the member id in a household, using 0 to index, 1 and so on as household contact, 3) size: the number of individuals in the households, 4) end: the end date of follow-up for that individual, 5) inf: the infection status of the member. By definition, the index must be infected. 6) onset: the onset time of the individual. The data must also include the factors affecting infectivity or susceptibility that may be explored.}

\item{inf_factor}{Factors affecting infectivity. Use the format of '~ factor1 + factor2'. If no factor, please use '~'}

\item{sus_factor}{Factors affecting susceptibility. Use the format of '~ factor1 + factor2'. If no factor, please use '~'}

\item{SI}{The mass function of the serial interval distribution. It should have a length equal to 14.}

\item{n_iteration}{The number of iterations of the MCMC.}

\item{burnin}{The iteration for burn-in for MCMC.}

\item{thinning}{The number of thinning in MCMC.}
}
\value{
A data frame that stores the estimates from a fitted MCMC.
}
\description{
The main function to fit the household transmission model to the data, to estimate the probability of infection from the community, probability of person-to-person transmission in households, and factors affecting susceptibility and infectivity.
}
\examples{
fitted_result <- household_dynamics(inputdata,'~sex','~age',SI,15000,5000,1,0)
}
