% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhdynamics_fit.R
\name{print.hhdynamics_fit}
\alias{print.hhdynamics_fit}
\title{Print method for hhdynamics_fit}
\usage{
\method{print}{hhdynamics_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hhdynamics_fit}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for hhdynamics_fit
}
