% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_function.R
\name{run_MCMC}
\alias{run_MCMC}
\title{Run the MCMC for the household transmission}
\usage{
run_MCMC(
  data_w,
  SI,
  n_iteration = 15000,
  burnin = 5000,
  thinning = 1,
  n_inf,
  n_sus,
  with_rm
)
}
\arguments{
\item{data_w}{The input data, in long format (each row is an individual). The data must include: 1) hhID: the household id, 2) member: the member id in a household, using 0 to index, 1 and so on as household contact, 3) size: the number of individuals in the households, 4) end: the end date of follow-up for that individual, 5) inf: the infection status of the member. By definition, the index must be infected. 6) onset: the onset time of the individual. The data must also include the factors affecting infectivity or susceptibility that may be explored.}

\item{SI}{The mass function of the serial interval distribution. It should have a length equal to 14.}

\item{n_iteration}{The number of iterations of the MCMC.}

\item{burnin}{The iteration for burn-in for MCMC.}

\item{thinning}{The number of thinning in MCMC.}

\item{n_inf}{The number of parameters affecting infectivity in the model.}

\item{n_sus}{The number of parameters affecting susceptibility in the model.}

\item{with_rm}{Indicator if the model has a random effect on individual infectivity or not.}
}
\value{
A matrix that stores the posterior samples for the model parameter.
}
\description{
This function runs the MCMC for the household transmission model. It is used in the main function.
}
\examples{
mcmc_result <- run_MCMC(data_w,SI,n_iteration = 15000,burnin = 5000,thinning = 1,n_inf,n_sus,with_rm)
}
